Moved background handling to callbacks
